fix otns status emission not working when built by cmake
