+++ examples/platforms/simulation/Makefile.am
@@ -54,7 +54,6 @@ PLATFORM_SOURCES                          = \
     system.c                                \
     uart.c                                  \
     virtual_time/alarm-sim.c                \
-    virtual_time/otns.c                     \
     virtual_time/platform-sim.c             \
     $(NULL)
